requirements.txt
telethon>=1.36.0
supabase>=2.3.0
openai>=1.30.0
neo4j>=5.20.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2
loguru>=0.7.0
